+++ tests/integ/modin/index/test_equals.py
@@ -23,22 +23,15 @@ from tests.integ.modin.sql_counter import SqlCounter, sql_count_checker
         ([1, 2, 3, 4], [1, 2, 3], False),  # extra value in left
     ],
 )
-@pytest.mark.parametrize("is_left_lazy", [True, False])
-@pytest.mark.parametrize("is_right_lazy", [True, False])
-def test_index_equals(lhs, rhs, expected, is_left_lazy, is_right_lazy):
-    query_count = int(is_left_lazy) + int(is_right_lazy)
-    with SqlCounter(query_count=query_count, join_count=query_count):
+def test_index_equals(lhs, rhs, expected):
+    with SqlCounter(query_count=2, join_count=2):
         native_result = native_pd.Index(lhs).equals(native_pd.Index(rhs))
         assert native_result == expected
 
-        snow_result = pd.Index(lhs, convert_to_lazy=is_right_lazy).equals(
-            pd.Index(rhs, convert_to_lazy=is_right_lazy)
-        )
+        snow_result = pd.Index(lhs).equals(pd.Index(rhs))
         assert snow_result == expected
 
-        mixed_result = pd.Index(lhs, convert_to_lazy=is_left_lazy).equals(
-            native_pd.Index(rhs)
-        )
+        mixed_result = pd.Index(lhs).equals(native_pd.Index(rhs))
         assert mixed_result == expected
 
 
@@ -60,16 +53,6 @@ def test_index_equals_other_types():
     index = pd.Index([1, 2])
     series = pd.Series([1, 2])
     assert index.equals(series) is False
-
-
-@sql_count_checker(query_count=2, join_count=2)
-def test_index_lazy_with_non_lazy():
-    # Lazy index
-    index1 = pd.Index([1, 2])
-    # Non lazy index
-    index2 = pd.Index([1, 2], convert_to_lazy=False)
-    assert index1.equals(index2)
-    assert index2.equals(index1)
 
 
 @sql_count_checker(query_count=0)